requirements.txt
requests
beautifulsoup4
python-dateutil
tqdm
pyyaml
# optional:
PyGithub
aiohttp
aiofiles
pandas
lxml
